Switch to matplotlib, add plotter utility
This change adds the IV_Swinger_plotter utility. It is a CLI tool that
allows the user to re-generate PDF plots from CSV data point files from
previous runs. It supports overlaying multiple curves on the same
graph, plotting the power curve with the IV curve, and many other
customization options. It can be run on a Mac or any other platform
that has the necessary Python support. Installing gnuplot on a Mac is a
bit dicey, so gnuplot has been replaced by matplotlib (pyplot).
Unfortunately on a gen 1 RPi B+ the matplotlib import adds several
seconds to the startup time, and the plotting and PDF generation takes
nearly 8 seconds. The plan is to restore gnuplot support in a
subsequent commit.

Other minor changes:
   - Interpolation is removed from first and last segments
     of the curve so they will always be linear on the graph
   - Option to plot ideal curve was removed