Restore gnuplot support
Importing matplotlib and using pyplot are too slow on a gen 1 RPi. This
commit adds gnuplot support back in to keep the performance of the
actual IV Swinger from degrading due to using matplotlib/pyplot. The
standalone plotter now has a —use_gnuplot option, but some of the
options are not supported when it is used.