Fix Info.plist (Mac)
The “Get Info” display for the Mac executable showed a version number
of 0.0.0.  This is the default generated by pyinstaller in the
Info.plist file in the app bundle. This commit adds a script,
fix_info_plist.py that is called by the mac_run_pyi script to correct
this. It also adds the copyright, and enables high resolution for
retina displays. The copyright string was also added to the “About”
dialog.