Bug fix: batch update was not saving updated configs correctly
When multiple runs were selected, and the “Update” button pressed, only
one run (the oldest one) was getting its saved .cfg file updated - and
it was getting updated with the newest one’s config values (which are
all identical, except for the title).  Needed one more line in the
update_selected method.