Calibration enhancements plus a few other things
The calibration values are now saved in Arduino EEPROM so each IVS2 box
can be independently calibrated; the laptop and IVS2 hardware are now
completely decoupled.  Added resistor values to the calibration.

Fixed default value for resistor R1 (180k -> 150k).

Also added a check of the Arduino sketch version at connection time and
a warning dialog if it is downlevel. The warning dialog has
instructions for updating it.

Added “View Config File” option to File menu.

Modified the change_folder method to do the right thing if the user
chooses the parent or child of the IV_Swinger2 folder.

Arduino sketch changes:
   - Added support for host writing/dumping EEPROM via config messages
   - Changed all static strings to use program memory to free up
dynamic memory