Split IV_Swinger2 module into GUI and non-GUI modules
Pretty major code refactoring.  All of the GUI code is now in the
IV_Swinger2_gui module. This is 2/3 of the code that was previously in
the IV_Swinger2 module, but never really belonged there.  The remaining
1/3 of the code is limited to the IVS2 configuration, hardware control,
plotting, and logging.  The configuration is no longer embedded in the
IV_Swinger2 class, but is now implemented in its own class that can be
extended by the higher level (e.g. GUI) to add additional configuration
that is applicable at that level.

With the GUI now fully decoupled from the lower level code, it is now
possible to write other programs that interact with the IV Swinger 2.
For example, it would be easy to implement a command line interface
(CLI). Perhaps more useful, however, is the ability to write automation
scripts that get the human out of the loop completely, and can
potentially also interact with other devices in conjunction with the
IVS2.