Fix for "missing points" bug, add noise reduction
Finally found and fixed the bug that occasionally caused curves to have
a large “gap” near the beginning - it was inadequate handing of relay
bounce in the Arduino sketch. The sketch now has conditionally compiled
code to help debug this type of problem by capturing all of the
unfiltered points. Also fixed another bug in the Arduino sketch that
caused it to not detect the “done” condition for much longer than
necessary.

Also improved the “correction” of the ADC values. A noise reduction
algorithm is now implemented.

Also changed the defaults for a couple of the Arduino preferences:
   - Max IV points is changed from 275 to 140. This just spaces the
points out so the red circles don’t overlap (with the default point
size). This should also improve loop rate on slower computers
   - Min Isc ADC is changed from 100 to 10. The resolution will
obviously be awful if it really is this low, but there’s no reason to
disallow it