Bug fixes and refactoring
- Fixed bug causing curve deflection at Voc end. This is fixed in the
v_adj property in the IV_Swinger2 module. An attempt is made to
determine where the curve will intersect the V axis before being
adjusted to intersect at Voc. The previous code assumed that the curve
was nearly vertical at this point, so it worked poorly when it was not.
- Remove extrapolated Isc point if the first measured point’s voltage
is more than 20% of Voc. We really don’t know Isc in this case, so
plotting it was misleading. (Previous commit removed the connecting
line, but this removes the plotted point too.)
 - Fixed update_selected() method so that it only updates the plotting
config and doesn’t modify the others (especially the calibration values)
 - Fixed problem where GUI comes up in the wrong position on the screen
 - Refactor: moved appropriate code from calibrate_battery_bias()
method from IV_Swinger2_gui module to IV_Swinger2 module