Fix minor bug and typo in help dialog
This fixes a minor bug: if the IVS2 was not connected when the app was
started, and then was connected after the app was already up, the
calibration menu would not have any of the entries disabled.  The
Voltage and Current entries require an IV curve to have been swung in
order to be valid.  Also, the Resistors entry was coming up enabled
even if the IVS2 was not connected (and would give a misleading error
message if it was selected). It must be disabled until the IVS2 is
connected.

A typo in the Calibration Help dialog was also corrected.