Fix bug in Arduino sketch and bump sketch version to 1.3.1
Fixed bug in Arduino sketch where a Voc=0 run would cause all
subsequent runs to appear to time out polling for stable Isc. An
Arduino reset cleared this up. The Arduino is reset when any Arduino
Preference is changed, so that action would appear to “fix” the
problem. Restarting the app also cleared it up.