Log exceptions and generate dialog for user + other minor changes
When an exception is taken by the code run by mainloop(), Tkinter does
not crash the program; it just prints the exception info to stderr and
keeps going. That makes it completely silent when the app is started
from an icon (and even when it is started from the CLI, you may not
notice it). Thanks (as usual) to StackOverflow, there’s a good solution
to this, which is to override Tkinter’s report_callback_exception
method. The overridden method prints the exception info to stdout and
also to the log file. It also generates an error dialog requesting the
user to e-mail me the log file. Decided to keep the behavior of not
crashing the app, however.

There are a few other more minor changes in this commit:
   - Change the procedure for Current Calibration in the calibration
help dialog (use light switch)
   - Change most tkmsg.showinfo calls to tkmsg.showerror
   - Fix restore_master method so it doesn’t cause a redisplay.  This
was causing corruption of the config file when an exception was
detected while running the Results Wizard (the corruption happened when
the Wizard was closed).
  - Catch exception on call to send2trash() method