Add support for active-high relays (issue #29)
Modify Arduino setup() code to read EEPROM location containing flag that indicates whether relay is active high. If EEPROM value is not valid, assume active-low.

Also add support for config message to control the primary relay (for manual testing).

Add checkbutton on Arduino tab of Preferences to specify that the relay is active-high.  Write the EEPROM location with the appropriate value.