Turn on second relay regardless of dynamic calibration (issue #30)
Change the swing_loop() method in IV_Swinger2_gui.py to turn the second relay on if battery bias mode is active, regardless of whether dynamic bias battery calibration is enabled.

Change the apply_battery_bias() method in IV_Swinger2.py to  compensate for the Vref droop from the second relay even if dynamic calibration is not enabled.