Modification to issue #30 fix
The second relay is now turned on unconditionally in the swing_loop() method. This has no effect on module IVS2 hardware. On the cell-version hardware it allows the bias battery and PV to be connected to the binding posts in the same way, even if Battery bias is not selected on the Plotting tab of Preferences.  In that case, the curve will be the combo curve.