Tweak to fix for issue #28
The message that is displayed on the screen when an error is detected while looping isn't guaranteed to fit. This was found when testing the feature added for issue #28 on a Windows 7 laptop.

The fix is to specify a negative number as the font size. This is interpreted not as a height in pixels, rather than a point size. This is more portable than using a point size.