Fix problem with fix for issue #25
The commit for issue #25 broke the reporting of errors returned by the receive_data_from_arduino() call in swing_iv_curve(). It added a call to send_config_msgs_to_arduino() and used the rc variable to capture its return code. This overwrote the value of rc from the receive_data_from_arduino() call.

The fix is to use a different variable for the receive_data_from_arduino() call return code.