Back off modification to issue #30 fix
Revert the change made in the 3-Jul commit.

Turning the 2nd relay on causes Vref droop, which is only accounted for when battery bias is enabled.

The only benefit to turning it on when battery bias is not enabled was that it seemed a bit more natural that you would get the combo curve (rather than the bias battery alone curve) if you have it connected in the bias config.  But this will be described correctly in the User Guide.