Fix grayed out File menu entries (Issue #36)
There's no way to make the file choosing dialog non-modal, since it is a MacOS dialog.  So to fix this problem, we just recreate the menu bar after the dialog is closed.  To do this, it was necessary to keep track of which entries in the Calibrate menu are enabled and disabled.  This is done with properties in the master class.

This might have been a better way to fix Issue #12 (without making the dialogs non-modal), but we're not going to change that now.