SSR module version PCB, RevB
Fixed bugs in RevA of the SSR module version PCB:

Issue #40: Must drive SSR1 and SSR2 from separate Arduino pins. In the RevA design, there is a path from +5V to GND through the two SSRs, which can cause destruction of the SSRs and the bleed resistor.

Issue #41: Resistor R3 must be connected precisely to the shunt (+) end pad so it doesn't see the voltage at some point in the sea of copper that is the LOAD_CAP- net.  This resulted in current measurements that were too high, and depended on whether the current was flowing from SSR3 or from the load caps.

Issue #41: The path from the shunt (-) end pad to the screw terminal pads must be a stub off the GND plane so that when current is flowing, the GND plane doesn't see a voltage drop.  Also, the linkage between this stub and the GND plane must be made at the shunt (-) end pad so the voltage of the GND plane is equal to the voltage at the (-) end of the shunt.

Also swapped the pads for the bypass diodes for layout reasons related to making the current-carrying part of GND a stub.