Standard (EMR-based) PCB update to RevB
This is RevB of the standard IVS2 EMR-based module version PCB.  It is based on RevB of the SSR-based module version. The main purpose of this revision is to make the SSR and EMR versions as similar as possible, namely:

  - the connections to the binding posts are both on a screw-terminal block on the left end of the board (like SSR version)

  - pads are provided for the bypass diodes (like SSR version)

  - the current-carrying path from the shunt is now isolated in a stub off the ground plane (like SSR version)

  - R3 is now connected to the shunt with a dedicated trace (like SSR version)

  - Connections to EMR are from a second screw-terminal block that is located on the board where the SSRs are in the SSR version

Note that there is nothing "wrong", per se, with the RevA board. This just cleans a few things up and makes it much easier to compare it with the SSR version.