Arduino code support for split control of SSR1/SSR2 (Issue #40)
This commit has the Arduino code change required for the hardware change for Issue #40, which split the control of SSR1 and SSR2 onto separate Arduino output pins.   It also has the code to turn off SSR3 during Isc polling when the voltage has stopped changing.  This cleans up the beginning of the curve.