Add control for cell version SSRs (SSR4, SSR5, SSR6) - Issue #43
Added Arduino code to control SSR4 and SSR6 which are present in the SSR cell version.  Since the Arduino outputs used to control these SSRs are not connected to anything in the other versions, those versions are not affected by this code.  SSR5 is connected to the same pin (D4) that is used for the "2nd relay" in the EMR cell version.