Minor change to debug logging
Previously, the log_initial_debug_info method, which is called before the first run, logged the values of some derived properties related to the voltmeter and ammeter circuits. These properties change their values, however, when the configuration (which contains calibration values) is applied.  This change moves the logging of those property values to a new method called log_meter_debug_info, which is called when the config is saved. It also adds the v_sat and i_sat properties to what it logs.

This change uncovered the fact that the config was being saved redundantly by the GUI, so that was fixed too.