Add Vref calibration (Issue #48)
Add independent calibration of the ADC reference voltage (Vref). This is the +5V voltage from USB.  This addresses Issue #48.

This calibration is stored in the config, but not in EEPROM on the Arduino.

Old results that were generated before this feature was added are not affected. When they are regenerated using the Update button in the Results Wizard, the default value of 5.0 is assumed for Vref.

The default current and voltage calibration values are now set to 1.0.

This commit contains a couple other minor unrelated changes. The Calibration help dialog was missing info on pyranometer calibration, so that was added. There was also another very minor pyranometer-related bug that was fixed.