Bring Results Wizard up immediately, and populate in the background (Issue #51)
- Pull treeview size configuration out of the populate_tree method and put it in its own method, configure_tree_size()

- Call configure_tree_size() before calling populate_tree() in the treeview() method

- Also insert "WORKING..." in the tree before populating

- Modify populate_tree() to call the delete_all() method at the beginning (and remove calls to delete_all() that precede calls to populate_tree())

- Change the call to populate_tree() in the treeview() method to use "after" so it is non-blocking

The effect of this is that when the Results Wizard button is clicked, the Wizard comes up immediately with "WORKING..." in the treeview pane. When the treeview is populated, that replaces the "WORKING..." and everything is as before. If the directory contains a small number of runs, the user will never even see the "WORKING..." message because it will be replaced so quickly with the populated treeview.