Clean up Calibrate menu item enabling/disabling (Issue #53)
This is the fix for Issue #54, but it also changes how all of the other Calibrate menu item states (enabled/disabled) are controlled to make it much simpler and cleaner.

All of the xxx_cal_enabled properties have been removed. Those properties were turned on and off in various places in the code and used in the update_calibrate_menu method.

Now the update_calibrate_menu method infers the state of the menu items from other properties and attributes. This removes about 100 lines of code, is easier to understand, and more mistakeproof.

One of the relevant properties is the arduino_ready property in the IV_Swinger2 class. This property currently does not always accurately represent the state of connection with the Arduino. Namely, if the USB cable is unplugged, it doesn't change to False.  To partially mitigate this, the getter method for the property has been changed to check if the USB has been disconnected and return False if it has (and change the property value to False).

Another related change that is in this commit is that the USB Port menu now has an update_usb_port_menu method that is registered as the postcommand callback, just like the Caibrate and File menus.  This method creates the menu items each time the menu is used. If the serial ports have changed, it will display the new list. Previously this was accomplished in the attempt_arduino_handshake method by removing and recreatiing the whole menubar.