Better/complete workaround for Mac gray menu bug (Issue #55)
This is a better fix for Issue #12 and Issue #36 that:
   - Also fix all other cases where the bug could manifest itself (all tk dialogs)
   - Restores the Dialog class to be modal

The new workaround is to recreate the whole menubar whenever a modal window is closed. This is the same workaround used for Issue #36, but is now enclosed in a method mac_grayed_menu_workaround. Calls to that method are placed after every usage of a dialog. To reduce the clutter, most of the tk dialogs (tkMessageBox, tkSimpleDialog, etc) are enclosed in wrappers that include the workaround.

The generic Dialog class is now reverted to its original modal behavior. All the dialog classes derived from Dialog no longer need the constrain_master and restore_master methods, so those (and calls to them) have been removed.