Work around for Mac Mojave problem (Issue #57)
Added self.update_idletasks() to the mac_grayed_menu_workaround method.  This seems to fix the problem with tkSimpleDialog buttons when running the pyinstaller-generated app on Mojave. Why it works isn't clear, nor is it clear that it will continue to work in the future, or even that it will work for everyone.