Another attempt at fix for Mac Mojave missing buttons (Issue #57)
This backs off the previous attempt at fixing Issue #57 and instead uses a clone of tkSimpleDialog that has one line added. This line is a call to update_idletasks() in the initializer method of the Dialog class.

This is an ugly hack and hopefuly can be removed at some point.