Update step-by-step docs
The main change was to consolidate Step 1 (Understand hardware) and Step 2 (Choose Variant) in the PCB docs. Now there is a single Step 1: Understand HW / Choose Variant".  The original step-by-step doc now also has the same Step 1. This keeps all of the step numbers the same in the original Instructable.