Add MPP logging (Issue #59)
Added logging of the following values at the calculated MPP:

  - Amps
  - Volts
  - Ohms
  - Watts

The log message is in the same format as the MPP log message for IV Swinger 1. All values have 6 digits after the decimal point, but that should not imply that there is that much precision in the measurements or MPP calculation.