Fix flawed pyranometer irradiance adjustment (Issue #60)
Replace the hardcoded irradiance adjustment with a configurable polynomial calibration.

Polynomial curve: y = Ax2 + Bx
  x: temperature-scaled photodiode millivolts
 A: PYRANO_CAL_A (W/m2/mV2)
 B: PYRANO_CAL (W/m2/mV)
 y: irradiance
If A is 0, scaling is linear. Intercept is always zero.

The default value for PYRANO_CAL_A is 0.0, so the scaling is linear. The GUI doesn't support setting PYRANO_CAL_A in the calibration dialog, so the only way to change its value is by manually editting the config file. The GUI calibration dialog is the same as before, but the code does take into account non-zero values of PYRANO_CAL_A such that when the user enters the measured irradiance, that will be the irradiance after the calibration.

This change also adds the photodiode temperature coefficent to the config file, so the user can change it. This is also not supported by the pyranometer calibration dialog, so it needs to be done by manually editting the config file. The default value for this coefficient was changed from 0.124 to 0.16.