Check that folders and files can be added to app data dir (Issue #64)
Add method check_app_data_dir() and call it in __init__ of class GraphicalUserInterface just after creating IV_Swinger2 object.

This method attempts to create a directory under the parent directory and a file under the app data directory. If either of these fails, a fatal error dialog is displayed and the program is exited when the dialog is dismissed.