Fix handling of non-ASCII characters (Issue #66)
The list of serial ports is obtained from serial.tools.list_ports.comports().  Each member of this list is an object of class serial.tools.list_ports.ListPortInfo. That class has a __str__ method that is called when the object is referenced in a string context. It returns a string containing the device and description properties.  This method throws a UnicodeEncodeError exception, however, if either of those properties contains non-ascii characters.

The fix is to not use str(serial_port), but to use serial_port.device and serial_port.description and use them as unicode.  It turns out that serial_port.manufacturer is useful too.

The code causing the crash is in find_arduino_port, but there is also code in the apply_usb and usb_disconnected methods requiring the same change.

Also, the update_usb_port_menu method in IV_Swinger2_gui.py requires a similar modification.

Another related change made in this commit is to check for non-ascii characters in the new title that the user enters for a run or overlay (change_title method).  An error dialog is displayed informing the user that the title must be ascii.