Use app_data_dir instead of results_dir for shortcut (Issue #74)
Shortcut is to the app data folder, not the results directory. This
is a cosmetic change because the results directory is the app data
folder unless it is changed by clicking on the column header. But then
the shortcut button changes to the import button, so there's no case
where this could matter.  Still, it's wrong.

This is one of the many cosmetic problems reported by Issue #74.