Add catch-all handler for early exceptions (Issue #75)
A try/except is now added in main() around the creation of the GraphicalUserInterface object and the call to its run() method. A new handle_early_exception() global function is called in the "except" branch. This function creates a temporary file and writes the error message and stack trace to that file. It then opens that file using the system viewer.