Check Voc / Isc calibration value to prevent divide-by-zero exception (Issue #84)
Added checks in get_v_cal_value() and get_i_cal_value().

To prevent the exception, the check just needs to be that the Voc/Isc value is not equal to the intercept value. However, it would never make sense for the intercept to be greater than the Voc/Isc value either, so the code detects and reports an error if the Voc/Isc is less than or equal to the intercept value. The user will see a dialog describing the error condition and the calibration is not performed.