Fix pylint "invalid-name" (Issue #87)
For the most part, this commit creates the IV_Swinger2_pylint.rc file, which (so far) only has "good-names" set to a list of names that are being blessed despite their non-compliance with pylint's criteria. Most are short names with no underscore. But there are also all of the class names that begin with "IV_Swinger", which do not conform to PascalCase. I decided a long time ago that PascalCase didn't work well for these names.

There are a few names in IV_Swinger2_gui.py and Tooltip.py that did get changed to be compliant since there wasn't a good reason not to.