Fix pylint "protected-access" (Issue #87)
This warning is disabled in the whole GraphicalUserInterfaceProps class because it is just an extension of the GraphicalUserInterface class.

It is also disabled in the code to add the icon to the Windows title bar because that code is copy/paste from StackOverflow and I don't know if there's a way around it.