Fix pylint "too-many-ancestors" (Issue #87)
All of these are disabled. The complaints are all about classes that extend a ttk class. There's not much that can be done if the library class has too many ancestors for pylint's liking.  It could be argued that this warning should be globally disabled since there is just no chance that any of the IV Swinger classes would have this problem.