Fix for Issue #89 - remove spurious CR on message from host
For some reason (not totally understood yet), the DUMP_EEPROM and DO_SSR_CURR_CAL messages from the host contain a trailing carriage return character when the application is running on Windows.

This commit adds code to the Arduino sketch to remove this spurious character if it is there. This fixes Issue #89.

The version number is incremented to 1.3.10.