Changes for code coverage (Issue #96)
The .coveragerc_Mac file contains the coverage exclusion patterns. It needs to be copied to .coveragerc before coverage is run on a Mac. A .coveragerc_Win file may be commited at a later time.

The changes to the Python files are all related to code coverage. Some unreachable code was removed. Some end-of-line comments were added to assist with the exclusions. The largest number of excluded lines are for code that is specific to IV Swinger 1. Others are for difficult-to-test error cases or other cases that are difficult or impossible to test.

This commit does not include the regression test list that is needed to achieve 100%. Currently I am using a list app (Wunderlist) for that.