Add feature to open run folder (Issue #97)
Added "Open Run Folder" entry to the File menu. This uses the system file viewer to open the run folder. This brings up Finder (Mac) or Explorer (Windows), with the run folder selected. This allows the user to open the CSV files to see the data values, or to copy files, remove files, etc. This is especially useful for regression testing (code coverage). Some users may be interested in seeing the data values in their spreadsheet application but this is mostly a developer feature.