More coverage updates (Issue #96)
Minor changes based on another code coverage pass.

One very minor functional change is "slipped in" with this commit. Fixes a quirk in the "restore looping options at startup" feature. (First diff in IV_Swinger2_gui.py)

Another was to fix an exception that occurs when an Update is performed on a run that does not have an ADC CSV file (IVS1 run).

The change to Tooltip.py was a line length issue caught by flake8.