Use Arduino internal 1.1V bandgap to measure Vref (Issue #100)
See Issue #100 for background.

Thanks to Leland Mayne for this idea and for the initial code.

This change is implemented such that the new Arduino sketch will work ok with older application versions and older Arduino sketches will work ok with the new application code. In either of those cases, the behavior will be the same as prior to this commit.

Executive summary: The Arduino sketch reads the bandgap value when requested by the host software and also every time the host sends a "Go" message. The IV_Swinger2.py code now supports calibration of the bandgap voltage and saving that value in EEPROM. It also parses the message from the Arduino that contains the bandgap measurement on every curve swing and uses that to update the Vref value. The IV_Swinger2_gui.py code now calibrates the bandgap value when a Vref calibration is performed. It also saves the updated Vref value to the configuration after each curve.

Arduino sketch changes:
  - Version bumped to 1.4.2 (Leland used 1.4.1, so I skipped that)
  - Added functions: set_up_bandgap(), read_bandgap() and read_internal_adc()
  - Added call to set_up_bandgap() in setup()
  - Added support for new READ_BANDGAP command from host which triggers a call to read_bandgap()
  - Added call to read_bandgap() that occurs on every "Go" command before the curve is swung.
  - Added call to read_bandgap() in the do_ssr_curr_cal() function
The read_bandgap() function prints a message to the host containing the total ADC value and the number of iterations performed.

IV_Swinger2.py changes:
  - Added EEPROM slot to save/retrieve the bandgap_microvolts value
  - Added update_vref() method to Configuration class
  - Added bandgap_microvolts, bandgap_total_adc, bandgap_iterations and avg_bandgap_adc properties to IV_Swinger2 class
  - Added methods read_bandgap(), parse_bandgap_msg(), calibrate_bandgap() and set_vref_from_bandgap() to IV_Swinger2 class
  - Added processing of bandgap message from Arduino to receive_data_from_arduino() and request_ssr_adv_current_calibration_val()

IV_Swinger2_gui.py changes:
  - Added call to configuration update_vref() method to swing_loop() after calling swing_curve()
  - Modified get_vref_cal_value() method in MenuBar class to call the IV_Swinger2 class's new calibrate_bandgap() method and to update the EEPROM with the calibrated value
  - Updated the CalibrationHelpDialog class's text to reflect the new scheme