Partial solution to dialog minimization problem (Issue #101)
This "fix" just catches the minimization event and puts up an error dialog that explains the problem and that the app will probably crash as soon as it is closed. This is only for dialogs that are based on the "Dialog" class, such as the Preferences dialog and the help dialogs. The hope is that this will at least let the user know that they should not attempt to minimize dialogs. It would have been nice to just not have a minimize button on dialogs, but that doesn't appear to be possbile.