Enforce 1-second swing interval more strictly (Issue #103)
The 1-second swing interval is now decoupled from the timestamp used for the run directory and file names. Now the actual time of the previous run is compared to the current time and a spin loop waits until it is at least one second.

Also, the loop time is now a floating-point number so fractional values such as 1.3 seconds can be used.