Add simulator feature (Issue #104)
The primary change in this commit is the addition of the IV_Swinger2_sim.py module. This implements a full-blown simulator for IV Swinger 2 that models the hardware and Arduino code to generate and plot predicted measured points for a given Isc, Voc and component values. It also can generate an optimal set of component values for a given target maximum Isc and Voc. The module also includes a graphical user interface dialog that can be used together with the main IV Swinger 2 application or used standalone.

The IV_Swinger2_gui.py is modified to add a "Run Simulator" entry to the Help menu, which brings up the simulator dialog. The decision to put this on the Help menu was because it didn't make sense on any of the other menus and didn't warrant its own menu since it is something that few users will need. It makes some sense to put it on the Help menu because its purpose is to help users choose components for scaled versions of the hardware.

The IV_Swinger2.py module has a very minor change to support showing the hardware limit values on the graph.