New fix for Issue #101
Back off previous "fix" for issue #101 which completely broke the tabs in the Preferences dialog.

This one actually fixes the problem - almost.  It turns out that removing the call to the resizable() method makes the problem disappear. By default tk.Toplevel widgets are resizable, so we just need to find a differen way to make non-resizable dialogs. This is done by setting both the minimum and maximum size to the current size.

Unfortunately, this solution has problems with the standalone version of the simulator dialog. But it works ok for the simulator dialog running under the IVS2 GUI.