Fix config file corruption cases (Issue #109)
The change to the Configuration class get() method in IV_Swinger2.py fixes the reordering problem. Instead of merging the file contents with current config, it blows away the old config and creates a new one.

There are several changes to IV_Swinger2_gui.py to fix the other issues. The main change is to not apply config changes to the old result that is still displayed after the Results Wizard is exited. For example, if the Plot Power checkbutton is checked immediately after exiting the Wizard, the plot that was being viewed is not updated to plot its power. This makes sense; if the user wants to add the power plot to an old run, that should be done while still in the Wizard. Checking plot power after the Wizard is exited changes the current config, not the config of the old run. There are other changes that were needed to fix the failing cases listed in Issue #109. The swap_config() and restore_config() methods were added to reduce code duplication in the GraphicalUserInterface.apply_new_ranges() and the
PlotPower.update_plot_power() methods.