Major enhancement: PV modeling and reference curves
This commit adds a major new module: IV_Swinger_PV_model.py. This is a general-purpose module that has no dependencies on any other IV Swinger modules, making it potentially useful for other projects. Its primary feature is the PV_model class that models a PV cell or module, given its datasheet specifications. Methods are provided to generate the PV's single-diode model parameters at a given cell temperature and irradiance and to generate the IV curve at those conditions.

A minor new module, IV_Swinger2_PV_model.py is also added. This module adds IV Swinger 2 specific features to the IV_Swinger_PV_model.py module.

The existing IV_Swinger.py and IV_Swinger_plotter.py modules are enhanced to support plotting the reference curve. These modules are not involved in generating the reference curve points, which are contained in a CSV file passed to them, just like the measured curves. The plotting is very similar to an overlay.

The IV_Swinger2.py module is enhanced to instantiate an IV_Swinger2_PV_model object in the IV_Swinger2 class. The Configuration class is enhanced to add a "PV model" section with the PV modeling options and to add a "plot ref" option to the "Plotting" section. Methods are added to the IV_Swinger2 class to use the PV model to generate a reference curve and to add it to the plot. There are also methods to generate and plot test curves using the PV model.

The IV_Swinger2_gui.py module is enhanced to add a "Plot Reference" checkbutton on the main screen and to add a "PV Model" tab to the Preferences dialog. The "PV Model" tab allows the user to enter datasheet specifications for PV modules (or cells) and to select from the list of previously entered PV modules/cells.  Entries from the list may also be deleted. There are buttons and entry widgets to allow the user to test the PV modeling and generate IV curves at a specified irradiance and temperature. There are also controls to specify the cell temperature adjustment value, whether to average multiple temperature sensors, and whether to override measured values with estimated irradiance and/or temperature.