Improve PV name label
This is a modification to the code for Issue #111.

On Windows, setting the font of the pv_name_label was resulting in much too large text (even if the font size is not specified). The point of using Courier was to  make it stand out.

Changed to use a custom style. It is based on TButton, so it has a box around it. Also use red foreground for further hoghlight. When the label is the empty string, the default label style "TLabel" is used.  The default font and size are used to avoid the mysterious Windows problem.