Use // for integer division (Issue #117)
There are only a few cases in the code where "classic int division" is performed by the / operator. This commit changes those / operators to the // operator, which keeps the behavior the same between Python 2 and Python 3.

In IV_Swinger.py, a bug was found where integer division was incorrect, and the numerator needed to change from 1 to 1.0 for it to be correct for Python 2. And one other change was made to change 0 to 0.0, even though it doesn't really matter.

The finddiv.py script (included in old Python releases) was used to find the division operators in the code. The -Qwarn option was also used to flag cases at runtime where classic int division was being performed. Without doing a full regression, however, it is not guaranteed that all cases are executed. The lines identified by finddiv.py were manually analyzed, and the vast majority of them were determined to operate on floats. It is possible that this manual analysis wasn't perfect. If any were missed, the effect in Python 3 will either be an exception (if the division result is required to be an integer) or it will be slightly different behavior.