Tcl/Tk 8.6.10 compatibility / Mac dark mode (Issue #118)
Created new style for disabled go_button that has black text and use that when the button is disabled.

Force the tooltip foreground (text) to black.

Change the color of the PV name label to a teal color that is legible with both light and dark backgrounds.

Remove the "" parameter from the offending self.tree.insert.