Run 2to3 script on python3 .py files (Issue #121)
This is step 1 of the port to Python 3. The 2to3 script is run on all of the .py files in the python3 directory (which were all identical to their couterparts in the python directory before this commit).

Additional changes are needed, but this commit shows the diffs that are made by the 2to3 tool with no additional modifications.