Remove utf-8 encode/decode (Issue #121)
The 2to3 script does not remove the explicit utf-8 encoding and decoding that was needed for Python 2. This is not only unnecessary for Python 3, but str objects do not have encode or decode methods, so it causes exceptions.

This commit simply removes the encode("utf-8") and decode("utf-8") calls. This is enough to allow the application to come up under Python 3. Many more changes are needed, however.