Update fix_info_plist.py for Python 3 (Issue #121)
The fix_info_plist.py script is updated so that it will run with either Python 2 or Python 3. Strings are updated to use .format(). The deprecated readPlist() and writePlist() methods are replaced by load() and dump(), but only when the script is run with Python 3. Dark Mode is enabled, by setting  NSRequiresAquaSystemAppearance to False. Hopefully this won't break anything when the app is run on older MacOS versions - we'll see. Copyright is updated to 2020.