Add logging of Python version and imported package versions (Issue #121)
Issue #120 should have requested logging of more than the Tcl/Tk version. Now that there is Python 2 and Python 3 code, we at least should log which one is running. And while we're at it, the versions of the imported packages that have a __version__ attribute should also be logged. This commit makes those changes to both the Python 2 and Python 3 code. The Python 2 code also has a workaround for a Matplotlib crash on Mac (doesn't happen with system python, but does with 2.7.18).