Fix spurious ListBox IndexError (Issue #123)
Add code at the beginning of the pv_model_listbox_select_actions method in the PreferecesDialog class to return immediately if the focus is not the listbox widget.  Since this method is called for reasons other than the <<ListboxSelect>> event, the widget check is only performed if the method is called with an event parameter.

Identical changes are made in the Python 2 and Python 3 code.